[REFACTOR] #64 프로필 이미지 변경 리팩토링
