+++ app/src/main/java/com/comjeong/nomadworker/data/datasource/source/mypage/MyPageDataSource.kt
@@ -1,15 +1,11 @@
 package com.comjeong.nomadworker.data.datasource.source.mypage
 
-import com.comjeong.nomadworker.data.model.mypage.ProfileImageResponseData
 import com.comjeong.nomadworker.data.model.mypage.UserFeedDetailResponseData
 import com.comjeong.nomadworker.data.model.mypage.UserTotalFeedsWithInfoResponseData
-import okhttp3.MultipartBody
 
 interface MyPageDataSource {
 
     suspend fun getUserTotalFeedsWithInfo(userId: Long): UserTotalFeedsWithInfoResponseData
 
     suspend fun getUserFeedDetail(feedId: Long): UserFeedDetailResponseData
-
-    suspend fun uploadUserProfileImage(profileImage: MultipartBody.Part): ProfileImageResponseData
 }
