+++ app/src/main/java/com/comjeong/nomadworker/data/datasource/source/settings/SettingsLocalDataSource.kt
@@ -2,7 +2,8 @@ package com.comjeong.nomadworker.data.datasource.source.settings
 
 import com.comjeong.nomadworker.data.model.settings.NoticeLocalData
 
-interface SettingsDataSource {
+interface SettingsLocalDataSource {
 
     fun fetchNoticeData(): MutableList<NoticeLocalData>
+
 }
