+++ app/src/main/java/com/comjeong/nomadworker/di/DataSourceModule.kt
@@ -1,12 +1,13 @@
 package com.comjeong.nomadworker.di
 
-import com.comjeong.nomadworker.data.datasource.local.SettingsDataSourceImpl
+import com.comjeong.nomadworker.data.datasource.local.SettingsLocalDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.mypage.MyPageRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.feed.FeedsRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.home.HomeRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.place.PlaceDetailRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.place.PlaceRegionRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.search.PlaceSearchDataSourceImpl
+import com.comjeong.nomadworker.data.datasource.remote.settings.SettingsRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.signin.SignInRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.remote.signup.SignUpRemoteDataSourceImpl
 import com.comjeong.nomadworker.data.datasource.source.SignInDataSource
@@ -17,7 +18,8 @@ import com.comjeong.nomadworker.data.datasource.source.mypage.MyPageDataSource
 import com.comjeong.nomadworker.data.datasource.source.place.PlaceDetailDataSource
 import com.comjeong.nomadworker.data.datasource.source.place.PlaceRegionDataSource
 import com.comjeong.nomadworker.data.datasource.source.search.PlaceSearchDataSource
-import com.comjeong.nomadworker.data.datasource.source.settings.SettingsDataSource
+import com.comjeong.nomadworker.data.datasource.source.settings.SettingsLocalDataSource
+import com.comjeong.nomadworker.data.datasource.source.settings.SettingsRemoteDataSource
 import org.koin.dsl.module
 
 val dataSourceModule = module {
@@ -29,5 +31,6 @@ val dataSourceModule = module {
     single<FeedsDataSource> { FeedsRemoteDataSourceImpl(get()) }
     single<MyPageDataSource> { MyPageRemoteDataSourceImpl(get(), get()) }
     single<PlaceSearchDataSource> { PlaceSearchDataSourceImpl(get()) }
-    single<SettingsDataSource> { SettingsDataSourceImpl() }
+    single<SettingsLocalDataSource> { SettingsLocalDataSourceImpl() }
+    single<SettingsRemoteDataSource> { SettingsRemoteDataSourceImpl(get()) }
 }
