+++ app/src/main/java/com/comjeong/nomadworker/di/RepositoryModule.kt
@@ -6,6 +6,7 @@ import com.comjeong.nomadworker.data.repository.mypage.MyPageRepositoryImpl
 import com.comjeong.nomadworker.data.repository.place.PlaceDetailRepositoryImpl
 import com.comjeong.nomadworker.data.repository.place.PlaceRegionRepositoryImpl
 import com.comjeong.nomadworker.data.repository.search.PlaceSearchRepositoryImpl
+import com.comjeong.nomadworker.data.repository.settings.SettingsRepositoryImpl
 import com.comjeong.nomadworker.data.repository.signin.SignInRepositoryImpl
 import com.comjeong.nomadworker.data.repository.signup.SignUpRepositoryImpl
 import com.comjeong.nomadworker.domain.repository.feed.FeedRepository
@@ -14,6 +15,7 @@ import com.comjeong.nomadworker.domain.repository.mypage.MyPageRepository
 import com.comjeong.nomadworker.domain.repository.place.PlaceDetailRepository
 import com.comjeong.nomadworker.domain.repository.place.PlaceRegionRepository
 import com.comjeong.nomadworker.domain.repository.search.PlaceSearchRepository
+import com.comjeong.nomadworker.domain.repository.settings.SettingsRepository
 import com.comjeong.nomadworker.domain.repository.signin.SignInRepository
 import com.comjeong.nomadworker.domain.repository.signup.SignUpRepository
 import org.koin.dsl.module
@@ -27,4 +29,5 @@ val repositoryModule = module {
     single<FeedRepository> { FeedRepositoryImpl(get()) }
     single<MyPageRepository> { MyPageRepositoryImpl(get()) }
     single<PlaceSearchRepository> { PlaceSearchRepositoryImpl(get()) }
+    single<SettingsRepository> { SettingsRepositoryImpl(get()) }
 }
