+++ app/src/main/java/com/comjeong/nomadworker/domain/repository/settings/SettingsRepository.kt
@@ -1,0 +1,9 @@
+package com.comjeong.nomadworker.domain.repository.settings
+
+import com.comjeong.nomadworker.domain.model.mypage.ProfileImageResult
+import okhttp3.MultipartBody
+
+interface SettingsRepository {
+
+    suspend fun updateUserProfileImage(profileImage: MultipartBody.Part): ProfileImageResult
+}
